Add null pointer check

+++ e2e/utils/test-utils.ts
@@ -505,6 +505,7 @@ export function parseForNumericalStrings(text: string): RegExpMatchArray | null 
 // Find and delete the workspace
 export async function cleanupWorkspace(page: Page, workspaceName: string) {
   const workspaceCard = await findWorkspaceCard(page, workspaceName);
-  await workspaceCard.delete();
-  return;
+  if (workspaceCard != null) {
+    await workspaceCard.delete();
+  }
 }
